feat(OrderMenu): 테이블 결제금액 계산 구현

+++ src/main/java/com/javabom/cafe/dto/order/OrderPayDto.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class OrderPayDto {
-    private int cash;
-    private int card;
+    private double cash;
+    private double card;
 }
